TEST: free translations must be valid IETF langauge tags
