+++ scription2dlx.js
@@ -58,4 +58,4 @@ t.exports=function(e){"use strict";if(!e.addUnicodeData)throw new ReferenceError
        * <xregexp.com>
        * Steven Levithan (c) 2007-2017 MIT License
        */
-"use strict";var n="xregexp",d={astral:!1,natives:!1},r={exec:RegExp.prototype.exec,test:RegExp.prototype.test,match:String.prototype.match,replace:String.prototype.replace,split:String.prototype.split},a={},f={},i={},c=[],o="default",s="class",l={default:/\\(?:0(?:[0-3][0-7]{0,2}|[4-7][0-7]?)?|[1-9]\d*|x[\dA-Fa-f]{2}|u(?:[\dA-Fa-f]{4}|{[\dA-Fa-f]+})|c[A-Za-z]|[\s\S])|\(\?(?:[:=!]|<[=!])|[?*+]\?|{\d+(?:,\d*)?}\??|[\s\S]/,class:/\\(?:[0-3][0-7]{0,2}|[4-7][0-7]?|x[\dA-Fa-f]{2}|u(?:[\dA-Fa-f]{4}|{[\dA-Fa-f]+})|c[A-Za-z]|[\s\S])|[\s\S]/},p=/\$(?:{([\w$]+)}|(\d\d?|[\s\S]))/g,h=void 0===r.exec.call(/()??/,"")[1],m=void 0!==/x/.flags,v={}.toString;function g(e){var t=!0;try{new RegExp("",e)}catch(e){t=!1}return t}var y=g("u"),b=g("y"),_={g:!0,i:!0,m:!0,u:y,y:b};function w(e,t,u,d,r){var a;if(e[n]={captureNames:t},r)return e;if(e.__proto__)e.__proto__=F.prototype;else for(a in F.prototype)e[a]=F.prototype[a];return e[n].source=u,e[n].flags=d?d.split("").sort().join(""):d,e}function E(e){return r.replace.call(e,/([\s\S])(?=[\s\S]*\1)/g,"")}function S(e,t){if(!F.isRegExp(e))throw new TypeError("Type RegExp expected");var u=e[n]||{},d=function(e){return m?e.flags:r.exec.call(/\/([a-z]*)$/i,RegExp.prototype.toString.call(e))[1]}(e),a="",f="",i=null,c=null;return(t=t||{}).removeG&&(f+="g"),t.removeY&&(f+="y"),f&&(d=r.replace.call(d,new RegExp("["+f+"]+","g"),"")),t.addG&&(a+="g"),t.addY&&(a+="y"),a&&(d=E(d+a)),t.isInternalOnly||(void 0!==u.source&&(i=u.source),null!=u.flags&&(c=a?E(u.flags+a):u.flags)),e=w(new RegExp(t.source||e.source,d),function(e){return!(!e[n]||!e[n].captureNames)}(e)?u.captureNames.slice(0):null,i,c,t.isInternalOnly)}function O(e){return parseInt(e,16)}function x(e,t,u){return"("===e.input.charAt(e.index-1)||")"===e.input.charAt(e.index+e[0].length)||function(e,t,u,n){var d=u.indexOf("x")>-1?["\\s","#[^#\\n]*","\\(\\?#[^)]*\\)"]:["\\(\\?#[^)]*\\)"];return r.test.call(new RegExp("^(?:"+d.join("|")+")*(?:"+n+")"),e.slice(t))}(e.input,e.index+e[0].length,u,"[?*+]|{\\d+(?:,\\d*)?}")?"":"(?:)"}function I(e){return parseInt(e,10).toString(16)}function A(e,t){var u,n=e.length;for(u=0;u<n;++u)if(e[u]===t)return u;return-1}function k(e,t){return v.call(e)==="[object "+t+"]"}function T(e){for(;e.length<4;)e="0"+e;return e}function M(e){var t={};return k(e,"String")?(F.forEach(e,/[^\s,]+/,function(e){t[e]=!0}),t):e}function L(e){if(!/^[\w$]$/.test(e))throw new Error("Flag must be a single character A-Za-z0-9_$");_[e]=!0}function P(e,t,u,n,d){for(var r,a,f=c.length,i=e.charAt(u),o=null;f--;)if(!((a=c[f]).leadChar&&a.leadChar!==i||a.scope!==n&&"all"!==a.scope||a.flag&&-1===t.indexOf(a.flag))&&(r=F.exec(e,a.regex,u,"sticky"))){o={matchLength:r[0].length,output:a.handler.call(d,r,n,t),reparse:a.reparse};break}return o}function C(e){d.astral=e}function N(e){RegExp.prototype.exec=(e?a:r).exec,RegExp.prototype.test=(e?a:r).test,String.prototype.match=(e?a:r).match,String.prototype.replace=(e?a:r).replace,String.prototype.split=(e?a:r).split,d.natives=e}function R(e){if(null==e)throw new TypeError("Cannot convert null or undefined to object");return e}function F(e,t){if(F.isRegExp(e)){if(void 0!==t)throw new TypeError("Cannot supply flags when copying a RegExp");return S(e)}if(e=void 0===e?"":String(e),t=void 0===t?"":String(t),F.isInstalled("astral")&&-1===t.indexOf("A")&&(t+="A"),i[e]||(i[e]={}),!i[e][t]){for(var u,n={hasNamedCapture:!1,captureNames:[]},d=o,a="",f=0,c=function(e,t){var u;if(E(t)!==t)throw new SyntaxError("Invalid duplicate regex flag "+t);for(e=r.replace.call(e,/^\(\?([\w$]+)\)/,function(e,u){if(r.test.call(/[gy]/,u))throw new SyntaxError("Cannot use flag g or y in mode modifier "+e);return t=E(t+u),""}),u=0;u<t.length;++u)if(!_[t.charAt(u)])throw new SyntaxError("Unknown regex flag "+t.charAt(u));return{pattern:e,flags:t}}(e,t),p=c.pattern,h=c.flags;f<p.length;){do{(u=P(p,h,f,d,n))&&u.reparse&&(p=p.slice(0,f)+u.output+p.slice(f+u.matchLength))}while(u&&u.reparse);if(u)a+=u.output,f+=u.matchLength||1;else{var m=F.exec(p,l[d],f,"sticky")[0];a+=m,f+=m.length,"["===m&&d===o?d=s:"]"===m&&d===s&&(d=o)}}i[e][t]={pattern:r.replace.call(a,/(?:\(\?:\))+/g,"(?:)"),flags:r.replace.call(h,/[^gimuy]+/g,""),captures:n.hasNamedCapture?n.captureNames:null}}var v=i[e][t];return w(new RegExp(v.pattern,v.flags),v.captures,e,t)}F.prototype=new RegExp,F.version="3.2.0",F._clipDuplicates=E,F._hasNativeFlag=g,F._dec=O,F._hex=I,F._pad4=T,F.addToken=function(e,t,u){var n,d=(u=u||{}).optionalFlags;if(u.flag&&L(u.flag),d)for(d=r.split.call(d,""),n=0;n<d.length;++n)L(d[n]);c.push({regex:S(e,{addG:!0,addY:b,isInternalOnly:!0}),handler:t,scope:u.scope||o,flag:u.flag,reparse:u.reparse,leadChar:u.leadChar}),F.cache.flush("patterns")},F.cache=function(e,t){return f[e]||(f[e]={}),f[e][t]||(f[e][t]=F(e,t))},F.cache.flush=function(e){"patterns"===e?i={}:f={}},F.escape=function(e){return r.replace.call(R(e),/[-\[\]{}()*+?.,\\^$|#\s]/g,"\\$&")},F.exec=function(e,t,u,d){var r,f,i="g",c=!1,o=!1;return(c=b&&!!(d||t.sticky&&!1!==d))?i+="y":d&&(o=!0,i+="FakeY"),t[n]=t[n]||{},f=t[n][i]||(t[n][i]=S(t,{addG:!0,addY:c,source:o?t.source+"|()":void 0,removeY:!1===d,isInternalOnly:!0})),u=u||0,f.lastIndex=u,r=a.exec.call(f,e),o&&r&&""===r.pop()&&(r=null),t.global&&(t.lastIndex=r?f.lastIndex:0),r},F.forEach=function(e,t,u){for(var n,d=0,r=-1;n=F.exec(e,t,d);)u(n,++r,e,t),d=n.index+(n[0].length||1)},F.globalize=function(e){return S(e,{addG:!0})},F.install=function(e){e=M(e),!d.astral&&e.astral&&C(!0),!d.natives&&e.natives&&N(!0)},F.isInstalled=function(e){return!!d[e]},F.isRegExp=function(e){return"[object RegExp]"===v.call(e)},F.match=function(e,t,u){var d,a,f=t.global&&"one"!==u||"all"===u,i=(f?"g":"")+(t.sticky?"y":"")||"noGY";return t[n]=t[n]||{},a=t[n][i]||(t[n][i]=S(t,{addG:!!f,removeG:"one"===u,isInternalOnly:!0})),d=r.match.call(R(e),a),t.global&&(t.lastIndex="one"===u&&d?d.index+d[0].length:0),f?d||[]:d&&d[0]},F.matchChain=function(e,t){return function e(u,n){var d=t[n].regex?t[n]:{regex:t[n]},r=[];function a(e){if(d.backref){if(!(e.hasOwnProperty(d.backref)||+d.backref<e.length))throw new ReferenceError("Backreference to undefined group: "+d.backref);r.push(e[d.backref]||"")}else r.push(e[0])}for(var f=0;f<u.length;++f)F.forEach(u[f],d.regex,a);return n!==t.length-1&&r.length?e(r,n+1):r}([e],0)},F.replace=function(e,t,u,d){var r,f=F.isRegExp(t),i=t.global&&"one"!==d||"all"===d,c=(i?"g":"")+(t.sticky?"y":"")||"noGY",o=t;return f?(t[n]=t[n]||{},o=t[n][c]||(t[n][c]=S(t,{addG:!!i,removeG:"one"===d,isInternalOnly:!0}))):i&&(o=new RegExp(F.escape(String(t)),"g")),r=a.replace.call(R(e),o,u),f&&t.global&&(t.lastIndex=0),r},F.replaceEach=function(e,t){var u,n;for(u=0;u<t.length;++u)n=t[u],e=F.replace(e,n[0],n[1],n[2]);return e},F.split=function(e,t,u){return a.split.call(R(e),t,u)},F.test=function(e,t,u,n){return!!F.exec(e,t,u,n)},F.uninstall=function(e){e=M(e),d.astral&&e.astral&&C(!1),d.natives&&e.natives&&N(!1)},F.union=function(e,t,u){var d,a,f=(u=u||{}).conjunction||"or",i=0;function c(e,t,u){var n=a[i-d];if(t){if(++i,n)return"(?<"+n+">"}else if(u)return"\\"+(+u+d);return e}if(!k(e,"Array")||!e.length)throw new TypeError("Must provide a nonempty array of patterns to merge");for(var o,s=/(\()(?!\?)|\\([1-9]\d*)|\\[\s\S]|\[(?:[^\\\]]|\\[\s\S])*\]/g,l=[],p=0;p<e.length;++p)o=e[p],F.isRegExp(o)?(d=i,a=o[n]&&o[n].captureNames||[],l.push(r.replace.call(F(o.source).source,s,c))):l.push(F.escape(o));var h="none"===f?"":"|";return F(l.join(h),t)},a.exec=function(e){var t,u,d,a=this.lastIndex,f=r.exec.apply(this,arguments);if(f){if(!h&&f.length>1&&A(f,"")>-1&&(u=S(this,{removeG:!0,isInternalOnly:!0}),r.replace.call(String(e).slice(f.index),u,function(){var e,t=arguments.length;for(e=1;e<t-2;++e)void 0===arguments[e]&&(f[e]=void 0)})),this[n]&&this[n].captureNames)for(d=1;d<f.length;++d)(t=this[n].captureNames[d-1])&&(f[t]=f[d]);this.global&&!f[0].length&&this.lastIndex>f.index&&(this.lastIndex=f.index)}return this.global||(this.lastIndex=a),f},a.test=function(e){return!!a.exec.call(this,e)},a.match=function(e){var t;if(F.isRegExp(e)){if(e.global)return t=r.match.apply(this,arguments),e.lastIndex=0,t}else e=new RegExp(e);return a.exec.call(e,R(this))},a.replace=function(e,t){var u,d,a,f=F.isRegExp(e);return f?(e[n]&&(d=e[n].captureNames),u=e.lastIndex):e+="",a=k(t,"Function")?r.replace.call(String(this),e,function(){var u,n=arguments;if(d)for(n[0]=new String(n[0]),u=0;u<d.length;++u)d[u]&&(n[0][d[u]]=n[u+1]);return f&&e.global&&(e.lastIndex=n[n.length-2]+n[0].length),t.apply(void 0,n)}):r.replace.call(null==this?this:String(this),e,function(){var e=arguments;return r.replace.call(String(t),p,function(t,u,n){var r;if(u){if((r=+u)<=e.length-3)return e[r]||"";if((r=d?A(d,u):-1)<0)throw new SyntaxError("Backreference to undefined group "+t);return e[r+1]||""}if("$"===n)return"$";if("&"===n||0==+n)return e[0];if("`"===n)return e[e.length-1].slice(0,e[e.length-2]);if("'"===n)return e[e.length-1].slice(e[e.length-2]+e[0].length);if(n=+n,!isNaN(n)){if(n>e.length-3)throw new SyntaxError("Backreference to undefined group "+t);return e[n]||""}throw new SyntaxError("Invalid token "+t)})}),f&&(e.global?e.lastIndex=0:e.lastIndex=u),a},a.split=function(e,t){if(!F.isRegExp(e))return r.split.apply(this,arguments);var u,n=String(this),d=[],a=e.lastIndex,f=0;return t=(void 0===t?-1:t)>>>0,F.forEach(n,e,function(e){e.index+e[0].length>f&&(d.push(n.slice(f,e.index)),e.length>1&&e.index<n.length&&Array.prototype.push.apply(d,e.slice(1)),u=e[0].length,f=e.index+u)}),f===n.length?r.test.call(e,"")&&!u||d.push(""):d.push(n.slice(f)),e.lastIndex=a,d.length>t?d.slice(0,t):d},F.addToken(/\\([ABCE-RTUVXYZaeg-mopqyz]|c(?![A-Za-z])|u(?![\dA-Fa-f]{4}|{[\dA-Fa-f]+})|x(?![\dA-Fa-f]{2}))/,function(e,t){if("B"===e[1]&&t===o)return e[0];throw new SyntaxError("Invalid escape "+e[0])},{scope:"all",leadChar:"\\"}),F.addToken(/\\u{([\dA-Fa-f]+)}/,function(e,t,u){var n=O(e[1]);if(n>1114111)throw new SyntaxError("Invalid Unicode code point "+e[0]);if(n<=65535)return"\\u"+T(I(n));if(y&&u.indexOf("u")>-1)return e[0];throw new SyntaxError("Cannot use Unicode code point above \\u{FFFF} without flag u")},{scope:"all",leadChar:"\\"}),F.addToken(/\[(\^?)\]/,function(e){return e[1]?"[\\s\\S]":"\\b\\B"},{leadChar:"["}),F.addToken(/\(\?#[^)]*\)/,x,{leadChar:"("}),F.addToken(/\s+|#[^\n]*\n?/,x,{flag:"x"}),F.addToken(/\./,function(){return"[\\s\\S]"},{flag:"s",leadChar:"."}),F.addToken(/\\k<([\w$]+)>/,function(e){var t=isNaN(e[1])?A(this.captureNames,e[1])+1:+e[1],u=e.index+e[0].length;if(!t||t>this.captureNames.length)throw new SyntaxError("Backreference to undefined group "+e[0]);return"\\"+t+(u===e.input.length||isNaN(e.input.charAt(u))?"":"(?:)")},{leadChar:"\\"}),F.addToken(/\\(\d+)/,function(e,t){if(!(t===o&&/^[1-9]/.test(e[1])&&+e[1]<=this.captureNames.length)&&"0"!==e[1])throw new SyntaxError("Cannot use octal escape or backreference to undefined group "+e[0]);return e[0]},{scope:"all",leadChar:"\\"}),F.addToken(/\(\?P?<([\w$]+)>/,function(e){if(!isNaN(e[1]))throw new SyntaxError("Cannot use integer as capture name "+e[0]);if("length"===e[1]||"__proto__"===e[1])throw new SyntaxError("Cannot use reserved word as capture name "+e[0]);if(A(this.captureNames,e[1])>-1)throw new SyntaxError("Cannot use same name for multiple groups "+e[0]);return this.captureNames.push(e[1]),this.hasNamedCapture=!0,"("},{leadChar:"("}),F.addToken(/\((?!\?)/,function(e,t,u){return u.indexOf("n")>-1?"(?:":(this.captureNames.push(null),"(")},{optionalFlags:"n",leadChar:"("}),t.exports=F},{}]},{},[8])(8)},function(e,t,u){"use strict";u.r(t);var n=u(51);function d(e,t){d=function(e,t){return new i(e,t)};var u=r(RegExp),n=RegExp.prototype,a=new WeakMap;function i(e,t){var n=u.call(this,e);return a.set(n,t),n}function c(e,t){var u=a.get(t);return Object.keys(u).reduce(function(t,n){return t[n]=e[u[n]],t},Object.create(null))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&f(e,t)}(i,u),i.prototype.exec=function(e){var t=n.exec.call(this,e);return t&&(t.groups=c(t,this)),t},i.prototype[Symbol.replace]=function(e,t){if("string"==typeof t){var u=a.get(this);return n[Symbol.replace].call(this,e,t.replace(/\$<([^>]+)>/g,function(e,t){return"$"+u[t]}))}if("function"==typeof t){var d=this;return n[Symbol.replace].call(this,e,function(){var e=[];return e.push.apply(e,arguments),"object"!=typeof e[e.length-1]&&e.push(c(e,d)),t.apply(this,e)})}return n[Symbol.replace].call(this,e,t)},d.apply(this,arguments)}function r(e){var t="function"==typeof Map?new Map:void 0;return(r=function(e){if(null===e||(u=e,-1===Function.toString.call(u).indexOf("[native code]")))return e;var u;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return a(e,arguments,i(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),f(n,e)})(e)}function a(e,t,u){return(a=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}()?Reflect.construct:function(e,t,u){var n=[null];n.push.apply(n,t);var d=new(Function.bind.apply(e,n));return u&&f(d,u.prototype),d}).apply(null,arguments)}function f(e,t){return(f=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function i(e){return(i=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function c(e){try{const t=function(e){const t=d(/\x2D\x2D\x2D((?:[\0-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])+?)\x2D\x2D\x2D/g,{header:1}).exec(e);return t?t.groups.header.trim():null}(e);if(null===t)return{};const u=function(e){if(""===e)throw new TypeError("The metadata header must not be empty.");return n.parse(e)}(t);return function(e){if("string"==typeof e)throw new Error("The metadata header could not be parsed as a JavaScript Object.");const t=e.title,u=e.utterances;if(!t)throw new Error('The metadata header must have a "title" attribute.');if(u)throw new Error('The metadata header must not have an "utterances" attribute.')}(u),u}catch(t){throw t.message="[getHeader] ".concat(t.message,"\n\nError parsing the following header:\n\n").concat(e),t}}function o(e,t){o=function(e,t){return new r(e,t)};var u=s(RegExp),n=RegExp.prototype,d=new WeakMap;function r(e,t){var n=u.call(this,e);return d.set(n,t),n}function a(e,t){var u=d.get(t);return Object.keys(u).reduce(function(t,n){return t[n]=e[u[n]],t},Object.create(null))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&p(e,t)}(r,u),r.prototype.exec=function(e){var t=n.exec.call(this,e);return t&&(t.groups=a(t,this)),t},r.prototype[Symbol.replace]=function(e,t){if("string"==typeof t){var u=d.get(this);return n[Symbol.replace].call(this,e,t.replace(/\$<([^>]+)>/g,function(e,t){return"$"+u[t]}))}if("function"==typeof t){var r=this;return n[Symbol.replace].call(this,e,function(){var e=[];return e.push.apply(e,arguments),"object"!=typeof e[e.length-1]&&e.push(a(e,r)),t.apply(this,e)})}return n[Symbol.replace].call(this,e,t)},o.apply(this,arguments)}function s(e){var t="function"==typeof Map?new Map:void 0;return(s=function(e){if(null===e||(u=e,-1===Function.toString.call(u).indexOf("[native code]")))return e;var u;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return l(e,arguments,h(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),p(n,e)})(e)}function l(e,t,u){return(l=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}()?Reflect.construct:function(e,t,u){var n=[null];n.push.apply(n,t);var d=new(Function.bind.apply(e,n));return u&&p(d,u.prototype),d}).apply(null,arguments)}function p(e,t){return(p=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function h(e){return(h=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}const m=o(/^\\((?:[\0-\x08\x0E-\x1F!-\x9F\xA1-\u167F\u1681-\u1FFF\u200B-\u2027\u202A-\u202E\u2030-\u205E\u2060-\u2FFF\u3001-\uD7FF\uE000-\uFEFE\uFF00-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])+)(?:[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]|$)/,{code:1}),v=/^[\x2D0-9A-Za-z]+$/,g=/\r?\n/g;function y(e){const t=e.match(m);return t?t.groups.code:null}function b(e){return e.some(e=>"string"==typeof e)}function _(e){const t=b(e),u=e.every(e=>"string"==typeof e);if(t&&!u)throw new Error("If one line in an utterance has a backslash code, all lines in the utterance must have backslash codes.");const n=e.filter(Boolean);n.reduce((e,t)=>{const u=e.get(t)||0;return e.set(t,u+1),e},new Map).forEach((e,t)=>{if("n"!==t&&e>1)throw new Error("The ".concat(t," code appears more than once in the utterance. Each backslash code may only appear once."))}),n.forEach(e=>{if(!function(e){return v.test(e)}(e))throw new Error("The backslash code ".concat(e," is invalid. Only characters A-Z, a-z, 0-9, and hyphens are allowed. Diacritics are not permitted."))})}function w(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var u=[],n=!0,d=!1,r=void 0;try{for(var a,f=e[Symbol.iterator]();!(n=(a=f.next()).done)&&(u.push(a.value),!t||u.length!==t);n=!0);}catch(e){d=!0,r=e}finally{try{n||null==f.return||f.return()}finally{if(d)throw r}}return u}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}const E={lit:["[","]"],phon:["[","]"],txn:["/","/"]};function S(e,t,u){return u.startsWith(e)&&u.endsWith(t)?u.slice(1,u.length-1).trim():u}function O(e,t){try{const u=E[e];return"string"==typeof t?t=S(...u,t):Object.entries(t).forEach(e=>{let n=w(e,2),d=n[0],r=n[1];t[d]=S(...u,r)}),t}catch(e){throw e.message="[cleanBrackets] ".concat(e.message),e}}function x(e,t){try{return(t.find(t=>{return t.code===e})||{}).data}catch(e){throw e.message="[getLine] ".concat(e.message),e}}function I(e,t){try{const u=t.filter(t=>{return t.type===e});return u.length?1===u.length&&u[0].code===e?u[0].data:u.reduce((e,t)=>{let u=t.code,n=t.data,d=t.type;return e[u.replace("".concat(d,"-"),"")]=n,e},{}):null}catch(e){throw e.message="[getLineGroup] ".concat(e.message,"\n\n").concat(t),e}}var A=u(50);const k=["phon","sp"];var T=u(52);function M(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var u=[],n=!0,d=!1,r=void 0;try{for(var a,f=e[Symbol.iterator]();!(n=(a=f.next()).done)&&(u.push(a.value),!t||u.length!==t);n=!0);}catch(e){d=!0,r=e}finally{try{n||null==f.return||f.return()}finally{if(d)throw r}}return u}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function L(e){try{let t=I("lit",e);return t?(function(e){if("object"==typeof e){const t=T(),u=e=>t.test(e);Object.entries(e).forEach(e=>{let t=M(e,1)[0];if(!u(t))throw new Error("".concat(t," is not a valid IETF language tag."))})}}(t=O("lit",t)),t):null}catch(e){throw e.message="[parseLiteral] ".concat(e.message),e}}function P(e,t){P=function(e,t){return new r(e,t)};var u=C(RegExp),n=RegExp.prototype,d=new WeakMap;function r(e,t){var n=u.call(this,e);return d.set(n,t),n}function a(e,t){var u=d.get(t);return Object.keys(u).reduce(function(t,n){return t[n]=e[u[n]],t},Object.create(null))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&R(e,t)}(r,u),r.prototype.exec=function(e){var t=n.exec.call(this,e);return t&&(t.groups=a(t,this)),t},r.prototype[Symbol.replace]=function(e,t){if("string"==typeof t){var u=d.get(this);return n[Symbol.replace].call(this,e,t.replace(/\$<([^>]+)>/g,function(e,t){return"$"+u[t]}))}if("function"==typeof t){var r=this;return n[Symbol.replace].call(this,e,function(){var e=[];return e.push.apply(e,arguments),"object"!=typeof e[e.length-1]&&e.push(a(e,r)),t.apply(this,e)})}return n[Symbol.replace].call(this,e,t)},P.apply(this,arguments)}function C(e){var t="function"==typeof Map?new Map:void 0;return(C=function(e){if(null===e||(u=e,-1===Function.toString.call(u).indexOf("[native code]")))return e;var u;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return N(e,arguments,F(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),R(n,e)})(e)}function N(e,t,u){return(N=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}()?Reflect.construct:function(e,t,u){var n=[null];n.push.apply(n,t);var d=new(Function.bind.apply(e,n));return u&&R(d,u.prototype),d}).apply(null,arguments)}function R(e,t){return(R=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function F(e){return(F=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function B(e){try{const t=e.filter(e=>{return"n"===e.code});return t.length?t.map(e=>{return function(e){try{const t=P(/^(?:[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]*((?:[\0-\t\x0B\f\x0E-\u2027\u202A-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])+)[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]*:)?[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]*((?:[\0-\t\x0B\f\x0E-\u2027\u202A-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])+)[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]*$/,{meta:1,text:2}).exec(e).groups;return t.meta,{text:t.text}}catch(t){throw t.message="Error parsing note line: ".concat(t.message,"\n\nNote line data: ").concat(e),t}}(e.data)}):[]}catch(e){throw e.message="[parseNotes] ".concat(e.message),e}}const D=/^[0-9A-Za-z]+$/;function j(e){try{const u=x("sp",e);if(!u)return null;if(t=u,!D.test(t))throw new Error("".concat(u," is not a valid value for the speaker line."));return u}catch(e){throw e.message="[parseSpeaker] ".concat(e.message),e}var t}function U(e,t,u){return t in e?Object.defineProperty(e,t,{value:u,enumerable:!0,configurable:!0,writable:!0}):e[t]=u,e}function Y(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var u=[],n=!0,d=!1,r=void 0;try{for(var a,f=e[Symbol.iterator]();!(n=(a=f.next()).done)&&(u.push(a.value),!t||u.length!==t);n=!0);}catch(e){d=!0,r=e}finally{try{n||null==f.return||f.return()}finally{if(d)throw r}}return u}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function K(e,t){K=function(e,t){return new r(e,t)};var u=W(RegExp),n=RegExp.prototype,d=new WeakMap;function r(e,t){var n=u.call(this,e);return d.set(n,t),n}function a(e,t){var u=d.get(t);return Object.keys(u).reduce(function(t,n){return t[n]=e[u[n]],t},Object.create(null))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&J(e,t)}(r,u),r.prototype.exec=function(e){var t=n.exec.call(this,e);return t&&(t.groups=a(t,this)),t},r.prototype[Symbol.replace]=function(e,t){if("string"==typeof t){var u=d.get(this);return n[Symbol.replace].call(this,e,t.replace(/\$<([^>]+)>/g,function(e,t){return"$"+u[t]}))}if("function"==typeof t){var r=this;return n[Symbol.replace].call(this,e,function(){var e=[];return e.push.apply(e,arguments),"object"!=typeof e[e.length-1]&&e.push(a(e,r)),t.apply(this,e)})}return n[Symbol.replace].call(this,e,t)},K.apply(this,arguments)}function W(e){var t="function"==typeof Map?new Map:void 0;return(W=function(e){if(null===e||(u=e,-1===Function.toString.call(u).indexOf("[native code]")))return e;var u;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return $(e,arguments,G(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),J(n,e)})(e)}function $(e,t,u){return($=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}()?Reflect.construct:function(e,t,u){var n=[null];n.push.apply(n,t);var d=new(Function.bind.apply(e,n));return u&&J(d,u.prototype),d}).apply(null,arguments)}function J(e,t){return(J=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function G(e){return(G=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}const Q=K(/^\\(?:(?:[0-9A-Z_a-z]|\x2D)+)((?:[\0-\t\x0B\f\x0E-\u2027\u202A-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])*)$/,{lineData:1}),V=/\r?\n/g;function z(e,t){try{const u=e.split(V).map(e=>e.trim()).map((e,u)=>{const n=t[u]||"n",d=Y(n.split("-",1),1)[0],r=e.match(Q);return{code:n,data:(r?r.groups.lineData:e).trim(),type:d}});if(u.every(e=>{return""===e.data}))return null;const n=function(e){try{return e.reduce((t,u)=>{let n=u.code,d=u.data,r=u.type;return!A.includes(r)&&(t[r]=I(r,e)),k.includes(r)&&n.startsWith("".concat(r,"-"))&&(t[n]=d),t},{})}catch(e){throw e.message="[parseCustom] ".concat(e.message),e}}(u),d=L(u),r=B(u),a=function(e){try{let t=x("phon",e);return t?t=O("phon",t):null}catch(e){throw e.message="[parsePhonetic] ".concat(e.message),e}}(u),f=j(u),i=function(e){try{return I("trs",e)||null}catch(e){throw e.message="[parseTranscript] ".concat(e.message),e}}(u),c=function(e){try{let t=I("txn",e);return t?t=O("txn",t):null}catch(e){throw e.message="[parseTranscription] ".concat(e.message),e}}(u)||"",o=function(e){try{return I("tln",e)||null}catch(e){throw e.message="[parseTranslation] ".concat(e.message),e}}(u)||"";return function(e){for(var t=1;t<arguments.length;t++){var u=null!=arguments[t]?arguments[t]:{},n=Object.keys(u);"function"==typeof Object.getOwnPropertySymbols&&(n=n.concat(Object.getOwnPropertySymbols(u).filter(function(e){return Object.getOwnPropertyDescriptor(u,e).enumerable}))),n.forEach(function(t){U(e,t,u[t])})}return e}({},n,d?{literal:d}:{},r.length?{notes:r}:{},a?{phonetic:a}:{},f?{speaker:f}:{},i?{transcript:i}:{},{transcription:c,translation:o})}catch(t){throw t.message="[parseUtterance] ".concat(t.message,"\n\nUtterance text:\n\n").concat(e),t}}function H(e){try{const t=e.split(/\x2D\x2D\x2D/g).map(e=>e.trim()).pop(),u=function(){const e="".concat("[ \\t\\v\\u00a0\\u1680\\u2000-\\u200a\\u2028\\u2029\\u202f\\u205f\\u3000\\ufeff]","*\\r?\\n"),t="(?:".concat(e,"){2,}");return new RegExp(t,"gsu")}();if(!t)return[];const n=function(e){try{const t=e.split(g).map(e=>e.trim()),u=t.map(y);if(_(u),!b(u)){const e=t.length;if(2===e)return["txn","tln"];if(3===e)return["m","gl","tln"];if(e>=4)return["txn","m","gl","tln"].fill("n",4);throw new Error("Cannot infer an interlinear gloss schema for utterances with fewer than 2 lines.")}return u}catch(t){throw t.message="[getSchema] ".concat(t.message,"\n\nUtterance text:\n\n").concat(e),t}}(t.split(u)[0]),d=e=>z(e,n);return t.split(u).map(d).filter(Boolean)}catch(t){const u=250;throw t.message="[getUtterances] ".concat(t.message,"\n\nError parsing utterances for the following text:\n\n").concat(e.slice(0,u)),t}}function Z(e,t,u){return t in e?Object.defineProperty(e,t,{value:u,enumerable:!0,configurable:!0,writable:!0}):e[t]=u,e}function q(e){return""===e.trim()?{}:function(e){for(var t=1;t<arguments.length;t++){var u=null!=arguments[t]?arguments[t]:{},n=Object.keys(u);"function"==typeof Object.getOwnPropertySymbols&&(n=n.concat(Object.getOwnPropertySymbols(u).filter(function(e){return Object.getOwnPropertyDescriptor(u,e).enumerable}))),n.forEach(function(t){Z(e,t,u[t])})}return e}({},c(e),{utterances:H(e)})}u.d(t,"default",function(){return q})}]).default});
+"use strict";var n="xregexp",d={astral:!1,natives:!1},r={exec:RegExp.prototype.exec,test:RegExp.prototype.test,match:String.prototype.match,replace:String.prototype.replace,split:String.prototype.split},a={},f={},i={},c=[],o="default",s="class",l={default:/\\(?:0(?:[0-3][0-7]{0,2}|[4-7][0-7]?)?|[1-9]\d*|x[\dA-Fa-f]{2}|u(?:[\dA-Fa-f]{4}|{[\dA-Fa-f]+})|c[A-Za-z]|[\s\S])|\(\?(?:[:=!]|<[=!])|[?*+]\?|{\d+(?:,\d*)?}\??|[\s\S]/,class:/\\(?:[0-3][0-7]{0,2}|[4-7][0-7]?|x[\dA-Fa-f]{2}|u(?:[\dA-Fa-f]{4}|{[\dA-Fa-f]+})|c[A-Za-z]|[\s\S])|[\s\S]/},p=/\$(?:{([\w$]+)}|(\d\d?|[\s\S]))/g,h=void 0===r.exec.call(/()??/,"")[1],m=void 0!==/x/.flags,v={}.toString;function g(e){var t=!0;try{new RegExp("",e)}catch(e){t=!1}return t}var y=g("u"),b=g("y"),_={g:!0,i:!0,m:!0,u:y,y:b};function w(e,t,u,d,r){var a;if(e[n]={captureNames:t},r)return e;if(e.__proto__)e.__proto__=F.prototype;else for(a in F.prototype)e[a]=F.prototype[a];return e[n].source=u,e[n].flags=d?d.split("").sort().join(""):d,e}function E(e){return r.replace.call(e,/([\s\S])(?=[\s\S]*\1)/g,"")}function S(e,t){if(!F.isRegExp(e))throw new TypeError("Type RegExp expected");var u=e[n]||{},d=function(e){return m?e.flags:r.exec.call(/\/([a-z]*)$/i,RegExp.prototype.toString.call(e))[1]}(e),a="",f="",i=null,c=null;return(t=t||{}).removeG&&(f+="g"),t.removeY&&(f+="y"),f&&(d=r.replace.call(d,new RegExp("["+f+"]+","g"),"")),t.addG&&(a+="g"),t.addY&&(a+="y"),a&&(d=E(d+a)),t.isInternalOnly||(void 0!==u.source&&(i=u.source),null!=u.flags&&(c=a?E(u.flags+a):u.flags)),e=w(new RegExp(t.source||e.source,d),function(e){return!(!e[n]||!e[n].captureNames)}(e)?u.captureNames.slice(0):null,i,c,t.isInternalOnly)}function O(e){return parseInt(e,16)}function x(e,t,u){return"("===e.input.charAt(e.index-1)||")"===e.input.charAt(e.index+e[0].length)||function(e,t,u,n){var d=u.indexOf("x")>-1?["\\s","#[^#\\n]*","\\(\\?#[^)]*\\)"]:["\\(\\?#[^)]*\\)"];return r.test.call(new RegExp("^(?:"+d.join("|")+")*(?:"+n+")"),e.slice(t))}(e.input,e.index+e[0].length,u,"[?*+]|{\\d+(?:,\\d*)?}")?"":"(?:)"}function I(e){return parseInt(e,10).toString(16)}function A(e,t){var u,n=e.length;for(u=0;u<n;++u)if(e[u]===t)return u;return-1}function k(e,t){return v.call(e)==="[object "+t+"]"}function T(e){for(;e.length<4;)e="0"+e;return e}function M(e){var t={};return k(e,"String")?(F.forEach(e,/[^\s,]+/,function(e){t[e]=!0}),t):e}function L(e){if(!/^[\w$]$/.test(e))throw new Error("Flag must be a single character A-Za-z0-9_$");_[e]=!0}function P(e,t,u,n,d){for(var r,a,f=c.length,i=e.charAt(u),o=null;f--;)if(!((a=c[f]).leadChar&&a.leadChar!==i||a.scope!==n&&"all"!==a.scope||a.flag&&-1===t.indexOf(a.flag))&&(r=F.exec(e,a.regex,u,"sticky"))){o={matchLength:r[0].length,output:a.handler.call(d,r,n,t),reparse:a.reparse};break}return o}function C(e){d.astral=e}function N(e){RegExp.prototype.exec=(e?a:r).exec,RegExp.prototype.test=(e?a:r).test,String.prototype.match=(e?a:r).match,String.prototype.replace=(e?a:r).replace,String.prototype.split=(e?a:r).split,d.natives=e}function R(e){if(null==e)throw new TypeError("Cannot convert null or undefined to object");return e}function F(e,t){if(F.isRegExp(e)){if(void 0!==t)throw new TypeError("Cannot supply flags when copying a RegExp");return S(e)}if(e=void 0===e?"":String(e),t=void 0===t?"":String(t),F.isInstalled("astral")&&-1===t.indexOf("A")&&(t+="A"),i[e]||(i[e]={}),!i[e][t]){for(var u,n={hasNamedCapture:!1,captureNames:[]},d=o,a="",f=0,c=function(e,t){var u;if(E(t)!==t)throw new SyntaxError("Invalid duplicate regex flag "+t);for(e=r.replace.call(e,/^\(\?([\w$]+)\)/,function(e,u){if(r.test.call(/[gy]/,u))throw new SyntaxError("Cannot use flag g or y in mode modifier "+e);return t=E(t+u),""}),u=0;u<t.length;++u)if(!_[t.charAt(u)])throw new SyntaxError("Unknown regex flag "+t.charAt(u));return{pattern:e,flags:t}}(e,t),p=c.pattern,h=c.flags;f<p.length;){do{(u=P(p,h,f,d,n))&&u.reparse&&(p=p.slice(0,f)+u.output+p.slice(f+u.matchLength))}while(u&&u.reparse);if(u)a+=u.output,f+=u.matchLength||1;else{var m=F.exec(p,l[d],f,"sticky")[0];a+=m,f+=m.length,"["===m&&d===o?d=s:"]"===m&&d===s&&(d=o)}}i[e][t]={pattern:r.replace.call(a,/(?:\(\?:\))+/g,"(?:)"),flags:r.replace.call(h,/[^gimuy]+/g,""),captures:n.hasNamedCapture?n.captureNames:null}}var v=i[e][t];return w(new RegExp(v.pattern,v.flags),v.captures,e,t)}F.prototype=new RegExp,F.version="3.2.0",F._clipDuplicates=E,F._hasNativeFlag=g,F._dec=O,F._hex=I,F._pad4=T,F.addToken=function(e,t,u){var n,d=(u=u||{}).optionalFlags;if(u.flag&&L(u.flag),d)for(d=r.split.call(d,""),n=0;n<d.length;++n)L(d[n]);c.push({regex:S(e,{addG:!0,addY:b,isInternalOnly:!0}),handler:t,scope:u.scope||o,flag:u.flag,reparse:u.reparse,leadChar:u.leadChar}),F.cache.flush("patterns")},F.cache=function(e,t){return f[e]||(f[e]={}),f[e][t]||(f[e][t]=F(e,t))},F.cache.flush=function(e){"patterns"===e?i={}:f={}},F.escape=function(e){return r.replace.call(R(e),/[-\[\]{}()*+?.,\\^$|#\s]/g,"\\$&")},F.exec=function(e,t,u,d){var r,f,i="g",c=!1,o=!1;return(c=b&&!!(d||t.sticky&&!1!==d))?i+="y":d&&(o=!0,i+="FakeY"),t[n]=t[n]||{},f=t[n][i]||(t[n][i]=S(t,{addG:!0,addY:c,source:o?t.source+"|()":void 0,removeY:!1===d,isInternalOnly:!0})),u=u||0,f.lastIndex=u,r=a.exec.call(f,e),o&&r&&""===r.pop()&&(r=null),t.global&&(t.lastIndex=r?f.lastIndex:0),r},F.forEach=function(e,t,u){for(var n,d=0,r=-1;n=F.exec(e,t,d);)u(n,++r,e,t),d=n.index+(n[0].length||1)},F.globalize=function(e){return S(e,{addG:!0})},F.install=function(e){e=M(e),!d.astral&&e.astral&&C(!0),!d.natives&&e.natives&&N(!0)},F.isInstalled=function(e){return!!d[e]},F.isRegExp=function(e){return"[object RegExp]"===v.call(e)},F.match=function(e,t,u){var d,a,f=t.global&&"one"!==u||"all"===u,i=(f?"g":"")+(t.sticky?"y":"")||"noGY";return t[n]=t[n]||{},a=t[n][i]||(t[n][i]=S(t,{addG:!!f,removeG:"one"===u,isInternalOnly:!0})),d=r.match.call(R(e),a),t.global&&(t.lastIndex="one"===u&&d?d.index+d[0].length:0),f?d||[]:d&&d[0]},F.matchChain=function(e,t){return function e(u,n){var d=t[n].regex?t[n]:{regex:t[n]},r=[];function a(e){if(d.backref){if(!(e.hasOwnProperty(d.backref)||+d.backref<e.length))throw new ReferenceError("Backreference to undefined group: "+d.backref);r.push(e[d.backref]||"")}else r.push(e[0])}for(var f=0;f<u.length;++f)F.forEach(u[f],d.regex,a);return n!==t.length-1&&r.length?e(r,n+1):r}([e],0)},F.replace=function(e,t,u,d){var r,f=F.isRegExp(t),i=t.global&&"one"!==d||"all"===d,c=(i?"g":"")+(t.sticky?"y":"")||"noGY",o=t;return f?(t[n]=t[n]||{},o=t[n][c]||(t[n][c]=S(t,{addG:!!i,removeG:"one"===d,isInternalOnly:!0}))):i&&(o=new RegExp(F.escape(String(t)),"g")),r=a.replace.call(R(e),o,u),f&&t.global&&(t.lastIndex=0),r},F.replaceEach=function(e,t){var u,n;for(u=0;u<t.length;++u)n=t[u],e=F.replace(e,n[0],n[1],n[2]);return e},F.split=function(e,t,u){return a.split.call(R(e),t,u)},F.test=function(e,t,u,n){return!!F.exec(e,t,u,n)},F.uninstall=function(e){e=M(e),d.astral&&e.astral&&C(!1),d.natives&&e.natives&&N(!1)},F.union=function(e,t,u){var d,a,f=(u=u||{}).conjunction||"or",i=0;function c(e,t,u){var n=a[i-d];if(t){if(++i,n)return"(?<"+n+">"}else if(u)return"\\"+(+u+d);return e}if(!k(e,"Array")||!e.length)throw new TypeError("Must provide a nonempty array of patterns to merge");for(var o,s=/(\()(?!\?)|\\([1-9]\d*)|\\[\s\S]|\[(?:[^\\\]]|\\[\s\S])*\]/g,l=[],p=0;p<e.length;++p)o=e[p],F.isRegExp(o)?(d=i,a=o[n]&&o[n].captureNames||[],l.push(r.replace.call(F(o.source).source,s,c))):l.push(F.escape(o));var h="none"===f?"":"|";return F(l.join(h),t)},a.exec=function(e){var t,u,d,a=this.lastIndex,f=r.exec.apply(this,arguments);if(f){if(!h&&f.length>1&&A(f,"")>-1&&(u=S(this,{removeG:!0,isInternalOnly:!0}),r.replace.call(String(e).slice(f.index),u,function(){var e,t=arguments.length;for(e=1;e<t-2;++e)void 0===arguments[e]&&(f[e]=void 0)})),this[n]&&this[n].captureNames)for(d=1;d<f.length;++d)(t=this[n].captureNames[d-1])&&(f[t]=f[d]);this.global&&!f[0].length&&this.lastIndex>f.index&&(this.lastIndex=f.index)}return this.global||(this.lastIndex=a),f},a.test=function(e){return!!a.exec.call(this,e)},a.match=function(e){var t;if(F.isRegExp(e)){if(e.global)return t=r.match.apply(this,arguments),e.lastIndex=0,t}else e=new RegExp(e);return a.exec.call(e,R(this))},a.replace=function(e,t){var u,d,a,f=F.isRegExp(e);return f?(e[n]&&(d=e[n].captureNames),u=e.lastIndex):e+="",a=k(t,"Function")?r.replace.call(String(this),e,function(){var u,n=arguments;if(d)for(n[0]=new String(n[0]),u=0;u<d.length;++u)d[u]&&(n[0][d[u]]=n[u+1]);return f&&e.global&&(e.lastIndex=n[n.length-2]+n[0].length),t.apply(void 0,n)}):r.replace.call(null==this?this:String(this),e,function(){var e=arguments;return r.replace.call(String(t),p,function(t,u,n){var r;if(u){if((r=+u)<=e.length-3)return e[r]||"";if((r=d?A(d,u):-1)<0)throw new SyntaxError("Backreference to undefined group "+t);return e[r+1]||""}if("$"===n)return"$";if("&"===n||0==+n)return e[0];if("`"===n)return e[e.length-1].slice(0,e[e.length-2]);if("'"===n)return e[e.length-1].slice(e[e.length-2]+e[0].length);if(n=+n,!isNaN(n)){if(n>e.length-3)throw new SyntaxError("Backreference to undefined group "+t);return e[n]||""}throw new SyntaxError("Invalid token "+t)})}),f&&(e.global?e.lastIndex=0:e.lastIndex=u),a},a.split=function(e,t){if(!F.isRegExp(e))return r.split.apply(this,arguments);var u,n=String(this),d=[],a=e.lastIndex,f=0;return t=(void 0===t?-1:t)>>>0,F.forEach(n,e,function(e){e.index+e[0].length>f&&(d.push(n.slice(f,e.index)),e.length>1&&e.index<n.length&&Array.prototype.push.apply(d,e.slice(1)),u=e[0].length,f=e.index+u)}),f===n.length?r.test.call(e,"")&&!u||d.push(""):d.push(n.slice(f)),e.lastIndex=a,d.length>t?d.slice(0,t):d},F.addToken(/\\([ABCE-RTUVXYZaeg-mopqyz]|c(?![A-Za-z])|u(?![\dA-Fa-f]{4}|{[\dA-Fa-f]+})|x(?![\dA-Fa-f]{2}))/,function(e,t){if("B"===e[1]&&t===o)return e[0];throw new SyntaxError("Invalid escape "+e[0])},{scope:"all",leadChar:"\\"}),F.addToken(/\\u{([\dA-Fa-f]+)}/,function(e,t,u){var n=O(e[1]);if(n>1114111)throw new SyntaxError("Invalid Unicode code point "+e[0]);if(n<=65535)return"\\u"+T(I(n));if(y&&u.indexOf("u")>-1)return e[0];throw new SyntaxError("Cannot use Unicode code point above \\u{FFFF} without flag u")},{scope:"all",leadChar:"\\"}),F.addToken(/\[(\^?)\]/,function(e){return e[1]?"[\\s\\S]":"\\b\\B"},{leadChar:"["}),F.addToken(/\(\?#[^)]*\)/,x,{leadChar:"("}),F.addToken(/\s+|#[^\n]*\n?/,x,{flag:"x"}),F.addToken(/\./,function(){return"[\\s\\S]"},{flag:"s",leadChar:"."}),F.addToken(/\\k<([\w$]+)>/,function(e){var t=isNaN(e[1])?A(this.captureNames,e[1])+1:+e[1],u=e.index+e[0].length;if(!t||t>this.captureNames.length)throw new SyntaxError("Backreference to undefined group "+e[0]);return"\\"+t+(u===e.input.length||isNaN(e.input.charAt(u))?"":"(?:)")},{leadChar:"\\"}),F.addToken(/\\(\d+)/,function(e,t){if(!(t===o&&/^[1-9]/.test(e[1])&&+e[1]<=this.captureNames.length)&&"0"!==e[1])throw new SyntaxError("Cannot use octal escape or backreference to undefined group "+e[0]);return e[0]},{scope:"all",leadChar:"\\"}),F.addToken(/\(\?P?<([\w$]+)>/,function(e){if(!isNaN(e[1]))throw new SyntaxError("Cannot use integer as capture name "+e[0]);if("length"===e[1]||"__proto__"===e[1])throw new SyntaxError("Cannot use reserved word as capture name "+e[0]);if(A(this.captureNames,e[1])>-1)throw new SyntaxError("Cannot use same name for multiple groups "+e[0]);return this.captureNames.push(e[1]),this.hasNamedCapture=!0,"("},{leadChar:"("}),F.addToken(/\((?!\?)/,function(e,t,u){return u.indexOf("n")>-1?"(?:":(this.captureNames.push(null),"(")},{optionalFlags:"n",leadChar:"("}),t.exports=F},{}]},{},[8])(8)},function(e,t,u){"use strict";u.r(t);var n=u(51);function d(e,t){d=function(e,t){return new i(e,t)};var u=r(RegExp),n=RegExp.prototype,a=new WeakMap;function i(e,t){var n=u.call(this,e);return a.set(n,t),n}function c(e,t){var u=a.get(t);return Object.keys(u).reduce(function(t,n){return t[n]=e[u[n]],t},Object.create(null))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&f(e,t)}(i,u),i.prototype.exec=function(e){var t=n.exec.call(this,e);return t&&(t.groups=c(t,this)),t},i.prototype[Symbol.replace]=function(e,t){if("string"==typeof t){var u=a.get(this);return n[Symbol.replace].call(this,e,t.replace(/\$<([^>]+)>/g,function(e,t){return"$"+u[t]}))}if("function"==typeof t){var d=this;return n[Symbol.replace].call(this,e,function(){var e=[];return e.push.apply(e,arguments),"object"!=typeof e[e.length-1]&&e.push(c(e,d)),t.apply(this,e)})}return n[Symbol.replace].call(this,e,t)},d.apply(this,arguments)}function r(e){var t="function"==typeof Map?new Map:void 0;return(r=function(e){if(null===e||(u=e,-1===Function.toString.call(u).indexOf("[native code]")))return e;var u;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return a(e,arguments,i(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),f(n,e)})(e)}function a(e,t,u){return(a=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}()?Reflect.construct:function(e,t,u){var n=[null];n.push.apply(n,t);var d=new(Function.bind.apply(e,n));return u&&f(d,u.prototype),d}).apply(null,arguments)}function f(e,t){return(f=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function i(e){return(i=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function c(e){try{const t=function(e){const t=d(/\x2D\x2D\x2D((?:[\0-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])+?)\x2D\x2D\x2D/g,{header:1}).exec(e);return t?t.groups.header.trim():null}(e);if(null===t)return{};const u=function(e){if(""===e)throw new TypeError("The metadata header must not be empty.");return n.parse(e)}(t);return function(e){if("string"==typeof e)throw new Error("The metadata header could not be parsed as a JavaScript Object.");const t=e.title,u=e.utterances;if(!t)throw new Error('The metadata header must have a "title" attribute.');if(u)throw new Error('The metadata header must not have an "utterances" attribute.')}(u),u}catch(t){throw t.message="[getHeader] ".concat(t.message,"\n\nError parsing the following header:\n\n").concat(e),t}}function o(e,t){o=function(e,t){return new r(e,t)};var u=s(RegExp),n=RegExp.prototype,d=new WeakMap;function r(e,t){var n=u.call(this,e);return d.set(n,t),n}function a(e,t){var u=d.get(t);return Object.keys(u).reduce(function(t,n){return t[n]=e[u[n]],t},Object.create(null))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&p(e,t)}(r,u),r.prototype.exec=function(e){var t=n.exec.call(this,e);return t&&(t.groups=a(t,this)),t},r.prototype[Symbol.replace]=function(e,t){if("string"==typeof t){var u=d.get(this);return n[Symbol.replace].call(this,e,t.replace(/\$<([^>]+)>/g,function(e,t){return"$"+u[t]}))}if("function"==typeof t){var r=this;return n[Symbol.replace].call(this,e,function(){var e=[];return e.push.apply(e,arguments),"object"!=typeof e[e.length-1]&&e.push(a(e,r)),t.apply(this,e)})}return n[Symbol.replace].call(this,e,t)},o.apply(this,arguments)}function s(e){var t="function"==typeof Map?new Map:void 0;return(s=function(e){if(null===e||(u=e,-1===Function.toString.call(u).indexOf("[native code]")))return e;var u;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return l(e,arguments,h(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),p(n,e)})(e)}function l(e,t,u){return(l=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}()?Reflect.construct:function(e,t,u){var n=[null];n.push.apply(n,t);var d=new(Function.bind.apply(e,n));return u&&p(d,u.prototype),d}).apply(null,arguments)}function p(e,t){return(p=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function h(e){return(h=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}const m=o(/^\\((?:[\0-\x08\x0E-\x1F!-\x9F\xA1-\u167F\u1681-\u1FFF\u200B-\u2027\u202A-\u202E\u2030-\u205E\u2060-\u2FFF\u3001-\uD7FF\uE000-\uFEFE\uFF00-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])+)(?:[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]|$)/,{code:1}),v=/^[\x2D0-9A-Za-z]+$/,g=/\r?\n/g;function y(e){const t=e.match(m);return t?t.groups.code:null}function b(e){return e.some(e=>"string"==typeof e)}function _(e){const t=b(e),u=e.every(e=>"string"==typeof e);if(t&&!u)throw new Error("If one line in an utterance has a backslash code, all lines in the utterance must have backslash codes.");const n=e.filter(Boolean);n.reduce((e,t)=>{const u=e.get(t)||0;return e.set(t,u+1),e},new Map).forEach((e,t)=>{if("n"!==t&&e>1)throw new Error("The ".concat(t," code appears more than once in the utterance. Each backslash code may only appear once."))}),n.forEach(e=>{if(!function(e){return v.test(e)}(e))throw new Error("The backslash code ".concat(e," is invalid. Only characters A-Z, a-z, 0-9, and hyphens are allowed. Diacritics are not permitted."))})}function w(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var u=[],n=!0,d=!1,r=void 0;try{for(var a,f=e[Symbol.iterator]();!(n=(a=f.next()).done)&&(u.push(a.value),!t||u.length!==t);n=!0);}catch(e){d=!0,r=e}finally{try{n||null==f.return||f.return()}finally{if(d)throw r}}return u}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}const E={lit:["[","]"],phon:["[","]"],txn:["/","/"]};function S(e,t,u){return u.startsWith(e)&&u.endsWith(t)?u.slice(1,u.length-1).trim():u}function O(e,t){try{const u=E[e];return"string"==typeof t?t=S(...u,t):Object.entries(t).forEach(e=>{let n=w(e,2),d=n[0],r=n[1];t[d]=S(...u,r)}),t}catch(e){throw e.message="[cleanBrackets] ".concat(e.message),e}}function x(e,t){try{return(t.find(t=>{return t.code===e})||{}).data}catch(e){throw e.message="[getLine] ".concat(e.message),e}}function I(e,t){try{const u=t.filter(t=>{return t.type===e});return u.length?1===u.length&&u[0].code===e?u[0].data:u.reduce((e,t)=>{let u=t.code,n=t.data,d=t.type;return e[u.replace("".concat(d,"-"),"")]=n,e},{}):null}catch(e){throw e.message="[getLineGroup] ".concat(e.message,"\n\n").concat(t),e}}var A=u(52);function k(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var u=[],n=!0,d=!1,r=void 0;try{for(var a,f=e[Symbol.iterator]();!(n=(a=f.next()).done)&&(u.push(a.value),!t||u.length!==t);n=!0);}catch(e){d=!0,r=e}finally{try{n||null==f.return||f.return()}finally{if(d)throw r}}return u}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function T(e){if("object"==typeof e){const t=A(),u=e=>t.test(e);Object.entries(e).forEach(e=>{let t=k(e,1)[0];if(!u(t))throw new Error("".concat(t," is not a valid IETF language tag."))})}}var M=u(50);const L=["phon","sp"];function P(e,t){P=function(e,t){return new r(e,t)};var u=C(RegExp),n=RegExp.prototype,d=new WeakMap;function r(e,t){var n=u.call(this,e);return d.set(n,t),n}function a(e,t){var u=d.get(t);return Object.keys(u).reduce(function(t,n){return t[n]=e[u[n]],t},Object.create(null))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&R(e,t)}(r,u),r.prototype.exec=function(e){var t=n.exec.call(this,e);return t&&(t.groups=a(t,this)),t},r.prototype[Symbol.replace]=function(e,t){if("string"==typeof t){var u=d.get(this);return n[Symbol.replace].call(this,e,t.replace(/\$<([^>]+)>/g,function(e,t){return"$"+u[t]}))}if("function"==typeof t){var r=this;return n[Symbol.replace].call(this,e,function(){var e=[];return e.push.apply(e,arguments),"object"!=typeof e[e.length-1]&&e.push(a(e,r)),t.apply(this,e)})}return n[Symbol.replace].call(this,e,t)},P.apply(this,arguments)}function C(e){var t="function"==typeof Map?new Map:void 0;return(C=function(e){if(null===e||(u=e,-1===Function.toString.call(u).indexOf("[native code]")))return e;var u;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return N(e,arguments,F(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),R(n,e)})(e)}function N(e,t,u){return(N=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}()?Reflect.construct:function(e,t,u){var n=[null];n.push.apply(n,t);var d=new(Function.bind.apply(e,n));return u&&R(d,u.prototype),d}).apply(null,arguments)}function R(e,t){return(R=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function F(e){return(F=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function B(e){try{const t=e.filter(e=>{return"n"===e.code});return t.length?t.map(e=>{return function(e){try{const t=P(/^(?:[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]*((?:[\0-\t\x0B\f\x0E-\u2027\u202A-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])+)[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]*:)?[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]*((?:[\0-\t\x0B\f\x0E-\u2027\u202A-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])+)[\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]*$/,{meta:1,text:2}).exec(e).groups;return t.meta,{text:t.text}}catch(t){throw t.message="Error parsing note line: ".concat(t.message,"\n\nNote line data: ").concat(e),t}}(e.data)}):[]}catch(e){throw e.message="[parseNotes] ".concat(e.message),e}}const D=/^[0-9A-Za-z]+$/;function j(e){try{const u=x("sp",e);if(!u)return null;if(t=u,!D.test(t))throw new Error("".concat(u," is not a valid value for the speaker line."));return u}catch(e){throw e.message="[parseSpeaker] ".concat(e.message),e}var t}function U(e,t,u){return t in e?Object.defineProperty(e,t,{value:u,enumerable:!0,configurable:!0,writable:!0}):e[t]=u,e}function Y(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var u=[],n=!0,d=!1,r=void 0;try{for(var a,f=e[Symbol.iterator]();!(n=(a=f.next()).done)&&(u.push(a.value),!t||u.length!==t);n=!0);}catch(e){d=!0,r=e}finally{try{n||null==f.return||f.return()}finally{if(d)throw r}}return u}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function K(e,t){K=function(e,t){return new r(e,t)};var u=W(RegExp),n=RegExp.prototype,d=new WeakMap;function r(e,t){var n=u.call(this,e);return d.set(n,t),n}function a(e,t){var u=d.get(t);return Object.keys(u).reduce(function(t,n){return t[n]=e[u[n]],t},Object.create(null))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&J(e,t)}(r,u),r.prototype.exec=function(e){var t=n.exec.call(this,e);return t&&(t.groups=a(t,this)),t},r.prototype[Symbol.replace]=function(e,t){if("string"==typeof t){var u=d.get(this);return n[Symbol.replace].call(this,e,t.replace(/\$<([^>]+)>/g,function(e,t){return"$"+u[t]}))}if("function"==typeof t){var r=this;return n[Symbol.replace].call(this,e,function(){var e=[];return e.push.apply(e,arguments),"object"!=typeof e[e.length-1]&&e.push(a(e,r)),t.apply(this,e)})}return n[Symbol.replace].call(this,e,t)},K.apply(this,arguments)}function W(e){var t="function"==typeof Map?new Map:void 0;return(W=function(e){if(null===e||(u=e,-1===Function.toString.call(u).indexOf("[native code]")))return e;var u;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return $(e,arguments,G(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),J(n,e)})(e)}function $(e,t,u){return($=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}()?Reflect.construct:function(e,t,u){var n=[null];n.push.apply(n,t);var d=new(Function.bind.apply(e,n));return u&&J(d,u.prototype),d}).apply(null,arguments)}function J(e,t){return(J=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function G(e){return(G=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}const Q=K(/^\\(?:(?:[0-9A-Z_a-z]|\x2D)+)((?:[\0-\t\x0B\f\x0E-\u2027\u202A-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])*)$/,{lineData:1}),V=/\r?\n/g;function z(e,t){try{const u=e.split(V).map(e=>e.trim()).map((e,u)=>{const n=t[u]||"n",d=Y(n.split("-",1),1)[0],r=e.match(Q);return{code:n,data:(r?r.groups.lineData:e).trim(),type:d}});if(u.every(e=>{return""===e.data}))return null;const n=function(e){try{return e.reduce((t,u)=>{let n=u.code,d=u.data,r=u.type;return!M.includes(r)&&(t[r]=I(r,e)),L.includes(r)&&n.startsWith("".concat(r,"-"))&&(t[n]=d),t},{})}catch(e){throw e.message="[parseCustom] ".concat(e.message),e}}(u),d=function(e){try{let t=I("lit",e);return t?(T(t=O("lit",t)),t):null}catch(e){throw e.message="[parseLiteral] ".concat(e.message),e}}(u),r=B(u),a=function(e){try{let t=x("phon",e);return t?t=O("phon",t):null}catch(e){throw e.message="[parsePhonetic] ".concat(e.message),e}}(u),f=j(u),i=function(e){try{return I("trs",e)||null}catch(e){throw e.message="[parseTranscript] ".concat(e.message),e}}(u),c=function(e){try{let t=I("txn",e);return t?t=O("txn",t):null}catch(e){throw e.message="[parseTranscription] ".concat(e.message),e}}(u)||"",o=function(e){try{const t=I("tln",e);return t?(T(t),t):null}catch(e){throw e.message="[parseTranslation] ".concat(e.message),e}}(u)||"";return function(e){for(var t=1;t<arguments.length;t++){var u=null!=arguments[t]?arguments[t]:{},n=Object.keys(u);"function"==typeof Object.getOwnPropertySymbols&&(n=n.concat(Object.getOwnPropertySymbols(u).filter(function(e){return Object.getOwnPropertyDescriptor(u,e).enumerable}))),n.forEach(function(t){U(e,t,u[t])})}return e}({},n,d?{literal:d}:{},r.length?{notes:r}:{},a?{phonetic:a}:{},f?{speaker:f}:{},i?{transcript:i}:{},{transcription:c,translation:o})}catch(t){throw t.message="[parseUtterance] ".concat(t.message,"\n\nUtterance text:\n\n").concat(e),t}}function H(e){try{const t=e.split(/\x2D\x2D\x2D/g).map(e=>e.trim()).pop(),u=function(){const e="".concat("[ \\t\\v\\u00a0\\u1680\\u2000-\\u200a\\u2028\\u2029\\u202f\\u205f\\u3000\\ufeff]","*\\r?\\n"),t="(?:".concat(e,"){2,}");return new RegExp(t,"gsu")}();if(!t)return[];const n=function(e){try{const t=e.split(g).map(e=>e.trim()),u=t.map(y);if(_(u),!b(u)){const e=t.length;if(2===e)return["txn","tln"];if(3===e)return["m","gl","tln"];if(e>=4)return["txn","m","gl","tln"].fill("n",4);throw new Error("Cannot infer an interlinear gloss schema for utterances with fewer than 2 lines.")}return u}catch(t){throw t.message="[getSchema] ".concat(t.message,"\n\nUtterance text:\n\n").concat(e),t}}(t.split(u)[0]),d=e=>z(e,n);return t.split(u).map(d).filter(Boolean)}catch(t){const u=250;throw t.message="[getUtterances] ".concat(t.message,"\n\nError parsing utterances for the following text:\n\n").concat(e.slice(0,u)),t}}function Z(e,t,u){return t in e?Object.defineProperty(e,t,{value:u,enumerable:!0,configurable:!0,writable:!0}):e[t]=u,e}function q(e){return""===e.trim()?{}:function(e){for(var t=1;t<arguments.length;t++){var u=null!=arguments[t]?arguments[t]:{},n=Object.keys(u);"function"==typeof Object.getOwnPropertySymbols&&(n=n.concat(Object.getOwnPropertySymbols(u).filter(function(e){return Object.getOwnPropertyDescriptor(u,e).enumerable}))),n.forEach(function(t){Z(e,t,u[t])})}return e}({},c(e),{utterances:H(e)})}u.d(t,"default",function(){return q})}]).default});
