+++ src/utilities/index.mjs
@@ -1,9 +1,11 @@
-import cleanBrackets from './cleanBrackets.mjs';
-import getLine       from './getLine.mjs';
-import getLines      from './getLines.mjs';
+import cleanBrackets     from './cleanBrackets.mjs';
+import getLine           from './getLine.mjs';
+import getLines          from './getLines.mjs';
+import validateLanguages from './validateLanguages.mjs';
 
 export {
   cleanBrackets,
   getLine,
   getLines,
+  validateLanguages,
 };
